Добавлена работа с базой данных. Рефакторинг усеченных коды.
[+] Добавлена работа с базой данных mysql. Теперь найденные коды можно сохранять в ней.
[*] Рефакторинг усеченных кодов. Методы для блоковых кодов и BEAST обощены для работы с усеченными кодами.

+++ src/trellises/ConvCodeTrellisIterator.java
@@ -33,7 +33,7 @@ public class ConvCodeTrellisIterator implements ITrellisIterator {
 	}
 	
 	@Override
-	public int vertexIndex() {
+	public long vertexIndex() {
 		return vertexIndex;
 	}
