+++ src/trellises/TrellisIterator.java
@@ -61,7 +61,7 @@ public class TrellisIterator implements ITrellisIterator {
 	}
 
 	@Override
-	public int vertexIndex() {
+	public long vertexIndex() {
 		return vertexIndex;
 	}
